+++ packages/sdks/web-component/src/lib/descope-wc/BaseDescopeWc.ts
@@ -272,6 +272,11 @@ class BaseDescopeWc extends BaseClass {
     }
   }
 
+  get popupOrigin(): string | null {
+    return this.getAttribute('popup-origin');
+  }
+
+  // grouped getter/setter for customStorage
   get customStorage(): CustomStorage | undefined {
     return this.#customStorage;
   }