+++ packages/sdks/nextjs-sdk/CHANGELOG.md
@@ -2,6 +2,45 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.14.31](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.30...nextjs-sdk-0.14.31) (2025-10-23)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.38.0`
+* `react-sdk` updated to version `2.22.0`
+* `core-js-sdk` updated to version `2.52.0`
+* `web-component` updated to version `3.48.0`
+## [0.14.30](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.29...nextjs-sdk-0.14.30) (2025-10-22)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.37.2`
+* `react-sdk` updated to version `2.21.4`
+* `core-js-sdk` updated to version `2.51.0`
+* `web-component` updated to version `3.47.12`
+## [0.14.29](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.28...nextjs-sdk-0.14.29) (2025-10-21)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.21.3`
+* `web-component` updated to version `3.47.11`
+
+### Bug Fixes
+
+* issue 12311 - nextjs type definitions ([#1230](https://github.com/descope/descope-js/issues/1230)) ([3cde977](https://github.com/descope/descope-js/commit/3cde97751d52b43e6f65c5d21204d815aa2aa75f))
+
+## [0.14.28](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.27...nextjs-sdk-0.14.28) (2025-10-16)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.37.1`
+* `react-sdk` updated to version `2.21.2`
+* `web-component` updated to version `3.47.10`
+## [0.14.27](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.26...nextjs-sdk-0.14.27) (2025-10-13)
+
+### Dependency Updates
+
+* `react-sdk` updated to version `2.21.1`
 ## [0.14.26](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.25...nextjs-sdk-0.14.26) (2025-10-09)
 
 ## [0.14.25](https://github.com/descope/descope-js/compare/nextjs-sdk-0.14.24...nextjs-sdk-0.14.25) (2025-09-29)